♻️ :: Category 도메인 응집도 향상
