+++ notification-domain/src/main/kotlin/io/github/v1servicenotification/category/api/response/CategoryListResponse.kt
@@ -1,0 +1,6 @@
+package io.github.v1servicenotification.category.api.response
+
+class CategoryListResponse(
+    val categories: List<CategoryElement>
+)
+
