--- notification-domain/src/main/kotlin/io/github/v1servicenotification/category/queryCategory/api/QueryNotificationCategory.kt
@@ -1,7 +1,0 @@
-package io.github.v1servicenotification.category.queryCategory.api
-
-import io.github.v1servicenotification.category.queryCategory.api.dto.response.CategoryListResponse
-
-interface QueryNotificationCategory {
-    fun queryNotificationCategory(defaultActivated: Boolean): CategoryListResponse
-}
