--- notification-domain/src/main/kotlin/io/github/v1servicenotification/category/queryCategory/api/dto/response/CategoryListResponse.kt
@@ -1,6 +1,0 @@
-package io.github.v1servicenotification.category.queryCategory.api.dto.response
-
-class CategoryListResponse(
-    val categories: List<CategoryElement>
-)
-
