+++ notification-domain/src/main/kotlin/io/github/v1servicenotification/category/spi/UpdateCategoryRepositorySpi.kt
@@ -1,4 +1,4 @@
-package io.github.v1servicenotification.category.updateCategory.spi
+package io.github.v1servicenotification.category.spi
 
 import io.github.v1servicenotification.annotation.Spi
 import io.github.v1servicenotification.category.Category
