--- notification-domain/src/main/kotlin/io/github/v1servicenotification/category/updateCategory/api/CreateCategory.kt
@@ -1,5 +1,0 @@
-package io.github.v1servicenotification.category.updateCategory.api
-
-interface CreateCategory {
-    fun createCategory(name: String, destination: String, defaultActivated: Boolean)
-}
