--- notification-domain/src/main/kotlin/io/github/v1servicenotification/category/updateCategory/api/RemoveCategory.kt
@@ -1,7 +1,0 @@
-package io.github.v1servicenotification.category.updateCategory.api
-
-import java.util.UUID
-
-interface RemoveCategory {
-    fun removeCategory(categoryId: UUID)
-}
