+++ notification-domain/src/main/kotlin/io/github/v1servicenotification/setting/activeSetting/api/QueryActivatedCategory.kt
@@ -1,6 +1,6 @@
 package io.github.v1servicenotification.setting.activeSetting.api
 
-import io.github.v1servicenotification.category.queryCategory.api.dto.response.CategoryListResponse
+import io.github.v1servicenotification.category.api.response.CategoryListResponse
 import java.util.*
 
 interface QueryActivatedCategory {
