+++ notification-domain/src/main/kotlin/io/github/v1servicenotification/stubs/InMemoryDetailRepository.kt
@@ -1,7 +1,7 @@
 package io.github.v1servicenotification.stubs
 
 import io.github.v1servicenotification.category.Category
-import io.github.v1servicenotification.category.updateCategory.exception.CategoryNotFoundException
+import io.github.v1servicenotification.category.exception.CategoryNotFoundException
 import io.github.v1servicenotification.detail.Detail
 import io.github.v1servicenotification.detail.exception.NotificationDetailNotFoundException
 import io.github.v1servicenotification.detail.postDetail.spi.PostDetailRepositorySpi
