+++ notification-domain/src/test/kotlin/io/github/v1servicenotification/detail/PostGroupNotificationImplTest.kt
@@ -1,7 +1,7 @@
 package io.github.v1servicenotification.detail
 
 import io.github.v1servicenotification.category.Category
-import io.github.v1servicenotification.category.updateCategory.exception.CategoryNotFoundException
+import io.github.v1servicenotification.category.exception.CategoryNotFoundException
 import io.github.v1servicenotification.detail.postDetail.service.PostGroupNotificationImpl
 import io.github.v1servicenotification.stubs.*
 import org.assertj.core.api.Assertions.assertThat
