+++ notification-infrastructure/src/main/kotlin/io/github/v1servicenotification/domain/category/domain/repository/CustomCategoryRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package io.github.v1servicenotification.domain.category.domain.repository
 
 import io.github.v1servicenotification.category.Category
-import io.github.v1servicenotification.category.queryCategory.spi.QueryCategoryRepositorySpi
-import io.github.v1servicenotification.category.updateCategory.spi.UpdateCategoryRepositorySpi
+import io.github.v1servicenotification.category.spi.QueryCategoryRepositorySpi
+import io.github.v1servicenotification.category.spi.UpdateCategoryRepositorySpi
 import io.github.v1servicenotification.domain.category.exception.CategoryNotFoundException
 import io.github.v1servicenotification.domain.category.mapper.CategoryMapper
 import io.github.v1servicenotification.global.extension.findOne
